added unit-test coverage for page.jsx
