+++ src/applications/virtual-agent/components/FloatingBot.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Disclaimer from './Disclaimer';
+import ChatBox from './Chatbox';
+
+export default function FloatingBot() {
+  return (
+    <div className="vads-l-grid-container large-screen:vads-u-padding-x--0">
+      <div className="vads-l-row vads-u-margin-x--neg2p5 vads-u-margin-y--4">
+        <div className="vads-l-col--12 vads-u-padding-x--2p5 medium-screen:vads-l-col--5 small-desktop-screen:vads-l-col--7">
+          <Disclaimer />
+        </div>
+        <div className="vads-l-col--12 vads-u-padding-x--2p5 medium-screen:vads-l-col--7 small-desktop-screen:vads-l-col--5">
+          <ChatBox />
+        </div>
+      </div>
+    </div>
+  );
+}
